Extract Parts and TimeUnit types
