+++ src/index.ts
@@ -15,3 +15,9 @@ export * from './conversion.js';
  */
 
 export * from './utilities.js';
+
+/**
+ * Types
+ */
+
+export * from './types.js';
