+++ src/types/TimeUnit.ts
@@ -1,0 +1,1 @@
+export type TimeUnit = 'ns' | 'μs' | 'ms' | 's' | 'm' | 'h' | 'd' | 'w';
